requirements.txt
selenium
undetected-chromedriver
pillow
python-dotenv
google-generativeai